Finished db for IVehicle and ICustomer. Issues with IBookings .

+++ CarRental.Common/Interface/ICustomer.cs
@@ -3,7 +3,7 @@ namespace CarRental.Common.Interface;
 
 public interface ICustomer
 {
-    public int sSD { get; init; }
+    public long sSD { get; init; }
     public string firstName { get; init; }
     public string lastName { get; init; }
 }
